Fix bugs, improve readability and rendering performance
- Fix sleep(0.033) → SDL_Delay(33) (was sleeping 0 seconds)
- Fix Figure_add_subplot returning object id instead of array index
- Fix plot_insert_point missing return (UB)
- Fix Subplot_AddPlot missing return (UB)
- Fix Subplot_Update null check after deref
- Fix Subplot_Destroy to use destroy vtable instead of bare free
- Fix Figure_create: add SDL_Init and malloc null check
- Fix Plotter.h stray array token (compilation error)
- Hoist Frame_GetTransform out of per-point loop in Data2Pixel
- Fix float cast to double in PlotLayer scale computation
- Remove unused Point struct, MAX_PLOTS, pxlValues field
- Make get_max/get_min/render/Figure_render static
- Fix unit tests to match current API

+++ PlotLayer.c
@@ -50,10 +50,10 @@ void PlotLayer_update(PlotLayer* self){
 
         f.parent = &self->base.frame;
 
-        f.Hom.R.rxx = (float)self->base.width/(self->data->x_max_displayed_point-self->data->x_min_displayed_point);
+        f.Hom.R.rxx = (double)self->base.width  / (self->data->x_max_displayed_point - self->data->x_min_displayed_point);
         f.Hom.R.ryx = 0;
-        f.Hom.R.rxy = 0; 
-        f.Hom.R.ryy = (float)self->base.height/(self->data->y_min_displayed_point-self->data->y_max_displayed_point);
+        f.Hom.R.rxy = 0;
+        f.Hom.R.ryy = (double)self->base.height / (self->data->y_min_displayed_point - self->data->y_max_displayed_point);
 
         f.Hom.t.u = -f.Hom.R.rxx * self->data->x_min_displayed_point;
         f.Hom.t.v = -f.Hom.R.ryy * self->data->y_max_displayed_point;
